YOLO

+++ src/main/java/com/epam/nyz/backend/model/CalculateResult.java
@@ -5,20 +5,20 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class CalculateResult {
 
-    private Long result;
+    private Double result;
 
     public CalculateResult() {
     }
 
-    public CalculateResult(Long result) {
+    public CalculateResult(Double result) {
         this.result = result;
     }
 
-    public Long getResult() {
+    public Double getResult() {
         return result;
     }
 
-    public void setResult(Long result) {
+    public void setResult(Double result) {
         this.result = result;
     }
